frame-mjpeg: Compute frame sizes in size_t
in->width and in->height are uint32_t, so width * height * N was
unsigned 32-bit arithmetic that wraps, the same latent problem fixed in
src/frame.c. The commit above adds a third such call, so convert all
three at once.

+++ src/frame-mjpeg.c
@@ -184,7 +184,7 @@ uvc_error_t uvc_mjpeg2bgr(uvc_frame_t *in, uvc_frame_t *out) {
   if (in->frame_format != UVC_FRAME_FORMAT_MJPEG)
     return UVC_ERROR_INVALID_PARAM;
 
-  if (uvc_ensure_frame_size(out, in->width * in->height * 3) < 0)
+  if (uvc_ensure_frame_size(out, (size_t)in->width * in->height * 3) < 0)
     return UVC_ERROR_NO_MEM;
 
   out->width = in->width;
@@ -209,7 +209,7 @@ uvc_error_t uvc_mjpeg2rgb(uvc_frame_t *in, uvc_frame_t *out) {
   if (in->frame_format != UVC_FRAME_FORMAT_MJPEG)
     return UVC_ERROR_INVALID_PARAM;
 
-  if (uvc_ensure_frame_size(out, in->width * in->height * 3) < 0)
+  if (uvc_ensure_frame_size(out, (size_t)in->width * in->height * 3) < 0)
     return UVC_ERROR_NO_MEM;
 
   out->width = in->width;
@@ -234,7 +234,7 @@ uvc_error_t uvc_mjpeg2gray(uvc_frame_t *in, uvc_frame_t *out) {
   if (in->frame_format != UVC_FRAME_FORMAT_MJPEG)
     return UVC_ERROR_INVALID_PARAM;
 
-  if (uvc_ensure_frame_size(out, in->width * in->height) < 0)
+  if (uvc_ensure_frame_size(out, (size_t)in->width * in->height) < 0)
     return UVC_ERROR_NO_MEM;
 
   out->width = in->width;
